I have just fixed a problem with one of the matcd audio ioctls.  The CDIOREADTOCENTRYS call was ignoring both the start_track and data_len fields of the call and returning the whole table of contents regardless.
This writes all over the stack of programs like workman and xcdplayer
which read the contents one track at a time.  I have included a patch
which fixes it.

The patch affects ONLY the CDIOREADTOCENTRYS ioctl of the matcd driver
and will not affect any other part of the system, especially not the
probe for the matcd device.
Submitted by:	Doug Rabson <dfr@render.com>


git-svn-id: svn+ssh://svn.freebsd.org/base/releng/2.0.5@8942 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f