Merge profiling bug fix from main branch 1.35 -> 1.36.
This fixes a serious profiling bug:  All addresses >= 256K caused
an overflow, and had their times tallied in the wrong counters.

Bruce Evans has devised a more complete fix that addresses other
issues as well.  But for now, this simple fix is effective and
easy to understand, making it worthy of inclusion in 2.1.5.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/2.1@16788 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f