[I'm still at COMDEX, but this is an easy commit]
David and I talked for quite awhile today about how to differentiate
2.1.6 from the "service patched" 2.1.6, and while the packaging
can and should stay the same, we still need to know version information
in bug reports.  I'm therefore bumping the version to 2.1.6.1, though
I hope we won't need to refer to it often as such since the 2.1.6 CD will
be, in fact, 2.1.6.1 and I intend to also replace the FTP release with it.
2.1.6a and 2.1.7 were also debated and rejected.  This version number
is the only one we've precedent for (1.1.5.1).


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/2.1@19932 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f