MFC:
Change vm_map_growstack to acquire and hold a read lock (instead of a write
lock) until it actually needs to modify the vm_map.

Upgrading a map's lock to exclusive status should increment
the map's timestamp.  In general, whenever an exclusive lock is
acquired the timestamp should be incremented.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/3@44651 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f