MFC: Remove obsolete isp_dogactive tag.
Add 8 bits of volatile mailbox busy mask- this will be the bitmask
of output mailbox values we want to get back out of the chip once
a mailbox command is done. Add storage for the maximum number of
output mailbox registers to the softc.  Set ISP_MAX_LUNS to be off
of new isp_maxluns tag in softc.  Change our 'fabdev' tag to
'loggedin'.  Increase size (and add defines for) topology storage.

Roll minor version numbers.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@62546 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f