As per discussion with ru, back out the part of the previous revision that introduced an example for a large filesystem (the patch addressed other issues as well).
The impact of a block:frag ratio other than 8:1 has not been
sufficiently investigated for its recommendation on stable production
systems.

This change just tries to limit the exposure of potentially bad advice,
and will be followed up with a sensible replacement (or the original
text) after consultation with imp and dillon.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@86978 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f