Partial MFC of rev. 1.79.
The I/F renaming as performed by the SIOCGIFCONF ioctl() is now in
sync with current. The linux_ifname interface function has not been
removed (as done in 1.79) or changed (as done in 1.82) due to the
11th hour in the 4.5 release cycle. This is fine as the function is
not used at all ATM and thus will not break anything. An MFC will
happen shortly after 4.5 is safely out the door.

Approved by: re


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@89677 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f