Allow one to specify a STANDARD_EXEC_PREFIX (used both for GCC and BINUTILS bits).
Also with the above, remove misleading comment about 'STANDARD_EXEC_PREFIX'.
It is used now and is needed if one does not set 'MD_EXEC_PREFIX'.

Do not define a 'MD_EXEC_PREFIX'.  It is not needed, not used in the
cross case, and just ends up causing "/usr/libexec" being added to the
library search path.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@91655 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f