MFC 1.57:
A long, long time ago, msmith introduced vfs.root.mountfrom
kernel tunnable, which let users specify the root mount point
the exact way one does after booting the kernel.

Let's take this opportunity to document it...

This is an immediate MFC due to the fact the line is commented
anyway.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@95525 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f