Make sysinstall's desktop configuration menu install the real kde and gnome metaports, rather than the less-featureful kdebase and gnomecore.
The general consensus seems to be that sysinstall and the disc 1 package
list were pruned down due to space considerations.  Those don't appear
to be a factor; in fact, we've been shipping the metaports on
disc 1 since 4.6.  This commit gets sysinstall caught up.

Approved by:	re
Discussed with:	portmgr (will)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@104232 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f