Forced commit:
MFC:

Change __sF to __std{in,out,err}p in #defines for std{in,out,err}.
This should help us with our compatibility problems over the long
haul.  I hope that this is the right thing to do.

std{in,out,err} are no longer compile time constants.  -current has
had this for 14 months now, and other systems do this.  I've fixed all
the places in the -stable tree that assumed that they were compile
time constants.

Reviewed by: arch@


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@106690 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f