MFC: rev 1.15
"pkg_create incorrectly does not add trailing '\n' when it receives
either COMMENT or DESCR from the command line. When a port is
installed, one gets both +COMMENT and +DESCR files with a trailing
'\n' character. However, +COMMENT does not contain a trailing '\n'
when it is installed from a package due to this behavior of pkg_create.

 Therefore, make sure it behaves exactly the same regardless of
where it got its information; either command line or files. The
modified functions are used by pkg_create."

PR:		52097


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@116095 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f