Disable PSE at boot time for Pentium 4 processors.  On many such CPUs, data corruption will result otherwise.  It seems that the bug is aggravated by recent work related to PAE.
This change was originally based on rev 1.386, but now the test for
CPU type is correct and the display of the warning is postponed until
we have initialized the kernel message buffer (e.g. for dmesg).

peter@ believes that only certain revisions of the Pentium 4 processor
are affected.  However, we are unsure which so we choose to be safe by
disabling PSE on all Pentium 4s.

The test can be disabled entirely with NO_PSE_HACK.

Approved by:	re


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@120879 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f