There is a comma missing in the table initializing the pmap_prefault_pageorder array.  This has two effects:
1. The resulting bogus contents of the array thwarts part of the
optimization effect pmap_prefault() is supposed to have.

2. The resulting array is only 7 elements long (auto-sized), while
pmap_prefault() expects it to be the intended 8 elements.  So this
function in fact accesses memory beyond the end of the array.
Fortunately though, if the data at this location is out of bounds it
will be ignored.

This bug dates back more than 6 years.  It has been introduced in
revision 1.178.

Submitted by:	Uwe Doering <gemini@geminix.org>
PR:	67460


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@129985 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f