Raise the ipl to splstatclock() in donice() before calling resetpriority(). This fixes races with schedcpu() and schedclock().  It corresponds to sched locking before the sched_nice() call in -current.  Lock assertions made the races very obvious in -current.
Adjust the process's priority after calling resetpriority() as in
schedclock() to maintain the invariant that the current priority is
always _the_ process's user priority if it is a user priority.  This
is to support uncommitted assertions that the priority is correct on
return to user mode.  The priority is correct here (modulo bugs
elsewhere) but it needs to be set explicitly to satisfy the assertions.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@130909 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f