MFC: Makefile.inc1 1.361 bin/ed/Makefile 1.22 lib/Makefile 1.157 lib/libcrypt/Makefile 1.31 lib/libfetch/Makefile 1.41 lib/libpam/modules/modules.inc 1.17 libexec/Makefile 1.61 share/examples/etc/make.conf 1.215 share/man/man5/make.conf.5 1.73 usr.bin/Makefile 1.239 usr.bin/fetch/Makefile 1.14 usr.sbin/pkg_install/Makefile 1.14 usr.sbin/pkg_install/add/Makefile 1.17 usr.sbin/pkg_install/info/Makefile 1.17 usr.sbin/pkg_install/version/Makefile 1.12 usr.sbin/ppp/Makefile 1.95, 1.96 usr.sbin/pppd/Makefile 1.25 usr.sbin/sendmail/Makefile 1.40 usr.sbin/tcpdump/tcpdump/Makefile 1.36
Retire the useless NOSECURE knob.

The motivation for this MFC is that MAKE_KERBEROS4 and MAKE_KERBEROS5
are meant to be buildable by various makefiles when NOSECURE is set but
this was broken because Kerberos depends on e.g. des(3) from OpenSSL
which wasn't built when NOSECURE was defined. Fixing the NOSECURE knob
however would have meant to make it do the same as the NOCRYPT knob.
If you had been using the NOSECURE build knob please now use the NOCRYPT
knob instead.

Discussed with:	ru (some months ago)
Reviewed by:	ru (some months ago)
Approved by:	markm (telnet and secure bits, some months ago), kensmith (re)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@138803 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f