MFC of 1.61, 1.62 of tcp_syncache.c and 1.107 tcp_var.h.
The tcp syncache code was leaving the IPv6 flowlabel uninitialised
for the SYN|ACK packet and then letting in6_pcbconnect set the
flowlabel later. Arange for the syncache/syncookie code to set and
recall the flow label so that the flowlabel used for the SYN|ACK
is consistent. This is done by using some of the cookie (when tcp
cookies are enabeled) and by stashing the flowlabel in syncache.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/4@140300 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f