MFC: camcontrol.c: 1.53 camcontrol.8: 1.40
Add "report only" functionality to 'camcontrol format', so users can get a
report on the status of a format already running on a drive.

Fix status reporting for 'camcontrol format'.  This was broken in rev 1.34
of camcontrol.c, almost 4 years ago!

Submitted by:	joerg (most of the reportonly changes)
Approved by:	re (kensmith)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/5@144336 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f