Merge subr_prof.c:1.119, 1.120, 1.121 from HEAD to RELENG_6:
  Acquire Giant in uprintf() and tprintf() due to the non-MPSAFEty of
  the tty code invoked from these functions.

  Correct age-old comments about uprintf()/tprintf() sleeping: they
  will never sleep.

Unlike in the merge to RELENG_6, it is not necesary to acquire Giant
in the NFS code, as the NFS client runs entirely under Giant in 5.x.
Likewise, this is not a critical fix in 5.x as the vast majority of
relevant calls to uprintf() are in the VFS code, which also runs
with Giant in 5.x.  However, there are some exposed code paths that
make this MFC a good idea.

Much useful feedback from:	bde
Approved by:			re (scottl)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/5@150722 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f