Call if_clone_detach() after vlan_clone_destroy() loop, as if_clone_detach() is freeing ifc_units field, which is used in vlan_clone_destroy()->ifc_free_unit().
This fixes panic: ifc_free_unit: bit is already cleared

The fix introduces a small race in that a new interface could be created
between the vlan_clone_destroy() loop and the call to if_clone_detach().
The race is probably too hard to trigger to be worry about it.

This is not an MFC, because better fix was committed to HEAD.
We decided to use this fix, as it is a low-risk fix, which fits just fine
for 6.0 at this stage.

Discussed with:	brooks, thompsa
Approved by:	re (scottl)


git-svn-id: svn+ssh://svn.freebsd.org/base/releng/6.0@151706 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f