MFC:	sys/kern/kern_malloc.c	1.150
In realloc(9), determine size of the original block based on
UMA_SLAB_MALLOC flag.
In some circumstances (I observed it when I was doing a lot of reallocs)
UMA_SLAB_MALLOC can be set even if us_keg != NULL.

If this is the case we have wonderful, silent data corruption, because less
data is copied to the newly allocated region than should be.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@154471 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f