MFC r1.12 -- back out r1.7.2.2.
As rse@ agreed, "status" should stay in $extra_commands
if an rc.d script doesn't use $command and default methods
associated with it.  This is true since rc.subr r1.42.
With r1.42 merged to rc.subr on RELENG_6 (thanks Doug!),
it's right time to re-add "status" to $extra_commands in
this script.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@154656 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f