MFC: 1.34
Extend the scope of #ifndef LOCORE to also cover the prototype of
ktr_tracepoint() and the macros using it. This allows this header
to be included in .S files for obtaining the KTR_* class macros
directly and providing a default value for KTR_COMPILE in case it's
not specified in the kernel config file including defining it to 0
when not using 'options KTR' at all.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@154718 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f