Destroy the cloned interfaces before, not after, detaching from the cloner structure in the module unload handler. The cloner structure was likely to be destroyed on detachment, which led to panic later, when its destroyed mutex was locked or whatever.  It was just another "use after free" case.
Now it is theoretically possible to create a vlan interface in
between the loop over the interfaces and the call to if_clone_detach().
We have to live with it in 6-STABLE.  In CURRENT, the actions are
serialized properly in if_clone_detach(): remove cloner from the
global list, destroy the cloned interfaces on it, detach from the
cloner.

This fix doesn't apply to HEAD, in which all the work is done by
if_clone_detach(), so it goes immediately to RELENG_6.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@161123 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f