MFC revision 1.101 (cognet@):
A temporary fix that in case of pwd == NULL, do not call audit_logout()
which attempts to deference it.  This is not quite correct, as we should
audit the event even it is not attributable to a specific user.  For now,
just put the temporary fix in, so login(1) would not get signal 11 upon
the case that for instance, ^D at the Login: prompt without providing a
valid login before.i

Approved by:	re (rwatson)
PR:		bin/103873
Discussed with:	rwatson, csjp


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@163056 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f