MFC (1.2: add missing ldexpf()).
Previous history of ldexpf()'s existence in FreeBSD:

2005/03/07 s_scalbnf.c 1.8: first implemented (for all arches)
           i387 s_scalbnf.S 1.1: i386 MD implementation at same time
2005/04/16 amd64 s_scalbnf.S 1.1: missing here, and thus missing for amd64
2005/09/12 amd64 s_scalbnf.S 1.2: fixed in -current
2005/12/06 e_scalbnf.c 1.10: misimplemented for all arches (accidentally has
                             no effect except for amd64)
2006/07/05 e_scalbnf.c 1.11: backed out misimplementation
2006/07/05 e_scalbnf.c 1.12: attempt to fix log message in 1.11

The log message in e_scalbnf.c 1.12 still got the main detail wrong.
1.10 actually succeeded in misimplementing ldexpf() for amd64 only,
since for amd64 only there is no strong reference to hide the broken
weak reference.  For most arches, there is a strong reference in
s_scalbnf.c; for i386, there is a strong reference in s_scalbnf.S; for
amd64 there is supposed to be a strong reference in s_scalbnf.S but
it was missing.

So the problem mainly affected all versions of RELENG_6 on amd64.  -current
has ldexpf() and versions before 6.0 aren't supposed to have it.

Debugging completed by: Kael Fischer, Steve Kargl


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@165113 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f