MFC: synchronize libarchive with FreeBSD-CURRENT.  This includes the following:
 * License updates.
 * Fixes for FreeBSD/arm
 * New read_open_filename, read_open_memory
 * New write_open_filename, write_open_memory
 * Write-blocking cleanup (cperciva@)
 * Adjust API so read_open/write_open funcs only need public API
 * Portability improvements (casts, include guards, etc.)
 * Properly pad sparse archive entries with blocks of zero bytes (cperciva@)
 * Correct copy of data to file when very large reads are possible
 * Be more flexible about dir permissions to avoid some fixups
 * Correct skip handling for entries over 2G (cperciva@)
 * ISO9660: skip entries faster, support entries over 2G
 * Tar: correct position accounting when skipping large entries (cperciva@)
 * Support unbuffered write as a special case (blocksize=0) (cperciva@)
 * Documentation corrections, improvements
 * Spelling corrections (cperciva@)
 * Improve some internal error returns

A few of these changes break the API/ABI and are therefore conditional
on the version number.  FreeBSD 6 will continue to use libarchive 1.x,
which will continue to use the old API/ABI.  When all of these changes
are ready, the libarchive version number will be changed in -CURRENT
to enable the new API.  Hopefully, this will allow the actual source
to remain identical between FreeBSD 6 and 7; only the version
parameters in the Makefile will be different.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@166273 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f