MFC :
Fix for a bug where only one process (of multiple) blocked on
maxpages on a zone is woken up, with the rest never being woken up as
a result of the ZFLAG_FULL flag being cleared. Wakeup all such blocked
procsses instead. This change introduces a thundering herd, but since
this should be relatively infrequent, optimizing this (by introducing
a count of blocked processes, for example) may be premature.

Reviewd by: ups@


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@166633 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f