Disable the module building of the linux bits on alpha and remove them from the corresponding NOTES. [1]
There's no corresponding linux_base in the Ports Collection since years,
and no linux distribution which is up-to-date and suitable (security,
widespread, ...) for us. The only complain regarding linux & alpha
we got so far is that the linuxulator doesn't build anymore. This
complaint came from wilko -- the alhpa-part of re@ -- only. As the
alpha platform is dead (no support from the manufacturer / linux
distributions) and deprecated (removed in -current) it was decided to
disable the linux bits from the build instead of trying to fix the
problems on alpha.

The alpha parts in the source are not removed so far, in case there are
1-2 people which actually still make use of this and are interested to
use the linuxulator (and fix the problems). Additionally this prevents
regressions on i386/amd64 when removing the alpha bits in shared code
(the people which care about those code parts don't have enough time ATM
to do extensive testing). This decission may be revisited at a later
point in time.

Discussed with:	wilko [1]


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@172251 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f