MFC: rev 1.452
Fix a bug that will cause a process that calls the VT_WAITACTIVE ioctl
to become unkillable when that process is sent a termination signal.  The
process will sit in waitvt looping in the kernel, and chewing up all
available CPU until the system is rebooted.

Submitted by:	Jilles Tjoelker <jilles@stack.nl>
Reviewed by:	bde


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/6@172392 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f