MFC v1.243 of config.c and v1.48 of dmenu.c - fixes needed because sysinstall had made bad assumptions about the behavior of getenv(3) and the new getenv(3) doesn't behave the same as the old one did if you call it with getenv("amd_enable=YES").
It was noted during review that the use of strdup(3) here is wrong
and should be checked for NULL.  It's worth noting that's what I meant
when I said I conformed to sysinstall's style despite urges to do otherwise
in the original commit.  sysinstall is in desperate need of some major
rototilling (or being thrown out...).

Approved by:	re (rwatson)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@173251 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f