MFC rev 1.31:
When the state of the interface changes rapidly enough there is a race
where dhclient is in the process of exiting due to the link going down
when the link coming up causes devd to try and start a new one.  This
results is the link being up, but no dhclient running.

Work around this race by checking a second time after a one second delay
before refusing to start a dhclient instance due to one already being
running.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@177113 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f