Extend global UNIX domain socket rwlock coverage to include incrementing and decrementing unp_rights, which may otherwise be corrupted.  Be slightly more conversative in where we read unp_rights.
This relates to one of the symptoms reported in the noted PR, but may not
correct the actual high system time problem.  The reporter has confirmed
stability but not that the problem is eliminated.  However, this is a
useful fix to a clear locking bug.

Note that this is not an MFC as the UNIX domain socket garbage collector
has been replaced in 8.x.

PR:		127024
Reported by:	Anton Yuzhaninov <citrin at citrin dot ru>
Reviewed by:	kib
Approved by:	re (kib)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@182993 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f