MFC r183288,183411,183827,184382:
Expand DBDMA API to allow setting device-dependent control bits and allow
DBDMA registers to lie in a subregion of a resource.

Also import changes to the BMAC driver to handle these changes and
change the way we enable the BMAC cell in macio. Instead of calling the
macio's enable-enet word, which apparently does nothing on some machines,
open an OF instance of the ethernet controller. This fixes cold booting
from disk on my Blue & White G3.

Approved by:	re (gnn)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@184461 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f