MFC rev. 182995
We can't implicitly trust the hook on NGQF_FN/NGQF_FN2 processing in
ng_apply_item(). There are possible (and I have got one) use-after-free
class panics because of it.

If hook is specified, require it to be valid at the apply time. The only
exceptions are the internal ng_con_part2(), ng_con_part3() and
ng_rmhook_part2() functions which are specially made to work with invalid
hooks.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@186976 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f