MFC 185714 from /head
Add a -L option to wc(1), for finger compatibility with the GNU
wc utility.  The -L option can be used to report the length of
the longest line wc has seen in one or more files.  It is
disabled by default, and wc uses the standard `-lwc'.

Submitted by:   Sheldon Givens, sheldon at sigsegv.ca


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@187285 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f