Fix unp_gc so that it recognizes file descriptors that are currently in the process of being passed between processes as alive and does not try to garbage collect them.
The full description of the problem and a test
program to reproduce it can be found in PR 112554.

This fix was inspired by similar fixes in NetBSD and BSD/OS.
However, it does not apply to FreeBSD 8 and above -
when this code was rewritten and optimized, the bug
was fixed in a different way.  The test program in the
PR passes on 8-current with flying colors.

PR:		112554
Submitted by:	Spencer Minear
Reviewed by:	Mike Silbersack
Obtained from:	Secure Computing Corp
MFC after:	4 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@190299 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f