MFC ixgbe version 1.7.4 for FreeBSD 7.2, this is earlier than planned but coordinated with and approved by RE.
Most important reason for this was so that FreeBSD 7.2
will have native support for the 82599 controller which
is part of the Nehalem launch.

The driver now does pluggable optics, multispeed fiber,
and header split. I am adding this to the GENERIC kernel
only in the amd64 architecture since its where I expect
most use and others were untested.

Enjoy!

Approved by: re


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@191030 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f