MFC r189878: Fix two issues with bufdaemon, often causing the processes to hang in the "nbufkv" sleep.
Do not retry request for the new block from ffs_bufwrite() when write is done
from bufdaemon and there is a buffer shortage.

In getnewbuf(), help bufdaemon to flush dirty buffers owned by the vnode
locked by curthread.

For MFC, default value for sysctl vfs.flushbufqtarget is set to -1,
disabling the helpers. The TDP_BUFNEED flag value from HEAD conflicts
with TDP_UPCALLING KSE bit, so it is moved to the end of allocated bits.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@191813 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f