use new btxld when building world in older environment
This is a direct commit to the branch.
This is for the benefit of those who define LOADER_ZFS_SUPPORT but
should't hurt anybody else.
Build of zfsboot requires new feature of btxld that it obtained
(via MFC) in r192697: "merge 177933. Allow for a zero length 'loader'"
(btxld is a build tool and a cross-build tool)
Please note that __FreeBSD_version was bumped some time after r192697
for an unrelated change.

Reviewed by:	des
Approved by:	jhb (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@193040 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f