dtrace/amd64: remove duplicate definition of dtrace_gethrtime function
Primary (and enhanced) definition of this function is in
sys/cddl/dev/dtrace/amd64/dtrace_subr.c.

This is a direct commit to the branch, because this instance of the
function was added to the branch in rev 182231, which claimed to be
dtrace MFH, but in fact the code in question never existed in head.
And does not exist now.

Another reason for this de-duplication is that there is only one instance
of this function for i386 arch.

Briefly discussed with:	jhb
Silence from:		jb


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@195493 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f