MFC	r196152
A piece of code was added to install a host route when an IPv6 interface
address is configured with a /128 prefix. This is no longer necessary due
to r192011. In fact that code conflicts with r192011. This patch removes
the host route installation when detecting the /128 prefix, and instead
let the code added by r192011 to install the loopback route for that IPv6
interface address.

Approved by:	re


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@196154 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f