MFC of r192625, r192890, r194783, r196587:
r192625:
MFC of tzcode2009e:

Upgrade of the tzcode from 2004a to 2009e.

Changes are numerous, but include...

- New format of the output of zic, which supports both 32 and 64
  bit time_t formats.

- zdump on 64 bit platforms will actually produce some output instead
  of doing nothing for a looooooooong time.

- linux_base-fX, with X >= at least 8, will work without problems related
  to the local time again.

The original patch, based on the 2008e, has been running for a long
time on both my laptop and desktop machine and have been tested by
other people.

After the installation of this code and the running of zic(8), you
need to run tzsetup(8) again to install the new datafile.

r192890:

MFC of tzcode2009h

- Clarify the license for the tzcode: public domain

r194783:

Remove duplicate if-statement on gmt_is_set in gmtsub().

r196587:

MFC of tzcode2009k

zic.c:
    Do not end a binary file with a POSIX-style time zone string
    for locations that end up in permanent DST (thanks to Andreas
    Schwab).


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@196683 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f