MFC	197695
Previously, if an address alias is configured on an interface, and
this address alias has a prefix matching that of another address
configured on the same interface, then the ARP entry for the alias
is not deleted from the ARP table when that address alias is removed.
This patch fixes the aforementioned issue.

PR:		kern/139113
Reviewed by:	bz
Approved by:	re


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@197811 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f