MFC r196754:
Add NFSv4 ACL support to cp(1) and fix a few memory leaks.

Note that this changes error reporting behaviour somewhat - before,
no error was reported if ACL couldn't be copied because the target
filesystem doesn't support ACLs.  Now, it will be reported - of course,
only if there actually is an ACL to copy.

Reviewed by:	rwatson


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@203189 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f