MFC r203665:
Make sure that FTS_COMFOLLOW is not set when the -P option is in effect.
Otherwise the -i option will show the inode number of the referenced file
for symbolic links given on the command line. Similarly, the file color
was printed according to the link target in colorized output.

PR:		bin/102394


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@204448 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f