MFC r204903:
Place interrupt handling in a critical section and remove double
counting in incrementing the interrupt nesting level. This fixes a number
of bugs in which the interrupt thread could be preempted by an IPI,
indefinitely delaying acknowledgement of the interrupt to the PIC, causing
interrupt starvation and hangs.

Reported by:	linimon
Reviewed by:	marcel, jhb


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@205375 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f