MFC r202668 and r200806 [1]:
Add a new option, -P, which reverts newsyslog(8) to the old behavior,
which stops to proceed further, as it is possible that processes which
fails to create PID file get screwed by rotation.

----
Don't consider non-existence of a PID file an error, we should be able
to proceed anyway as this most likely mean that the process has been
terminated.

PR:             bin/140397
Submitted by:   Dan Lukes <dan obluda cz>


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@207587 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f