MFC of 207924:
This fixes a bug with the one-2-one model socket when a
user sets up a socket to a server sends data and closes
the socket before the server has called accept(). It used
to NOT work at all. Now we add a flag to the assoc and
defer assoc cleanup so that the accept will succeed


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@208154 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f