MFC r208152,208172:
On PowerMac11,2 and (presumably) PowerMac12,1, we need to quiesce the
firmware in order to take over control of the SMU. Without doing this,
the firmware background process doing fan control will run amok as we
take over the system and crash the management chip.

This is limited to these two machines because our kernel is heavily
dependent on firmware accesses, and so quiescing firmware can cause
nasty problems.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@208427 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f