MFC 208603,208605: More gracefully handle stale file handles and attributes when opening a file via NFS.  Specifically, to satisfy close-to-open-consistency, the NFS client always performs at least one RPC on a file during an open(2) to see if the file has changed.  Normally this RPC is an ACCESS or GETATTR RPC that is forced by flushing a file's attribute cache during nfs_open() and then requesting new attributes.  However, if the file is noticed to be stale during nfs_open(), the only recourse is to fail the open(2) call with ESTALE.  On the other hand, if the ACCESS or GETATTR RPC is sent during nfs_lookup(), then the NFS client can fall back to a LOOKUP RPC to obtain the new file handle in the case that a file has been replaced.
This change causes the NFS client to flush the attribute cache during
nfs_lookup() when validating a name cache hit if the attributes fetched
during nfs_lookup() can be reused in nfs_open().  This allows the client
to open a replaced file via the new file handle the first time that it
notices a replaced file rather than failing with ESTALE in some cases.

Approved by:	re (kib)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@209003 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f