MFC r210736: sh: Do not enter consecutive duplicates into the history.
This simply sets a flag in libedit. It has a shortcoming in that it does not
apply to multi-line commands.

Note that a configuration option for this is not going to happen, but always
having this seems better than not having it. NetBSD has done the same.

PR:		bin/54683
Obtained from:	NetBSD


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@212111 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f