MFC r209053:
Add INVARIANTS checking that numfreebufs values are sane.  Also add a
per-buf flag to catch if a buf is double-counted in the free count.
This code was useful to debug an instance where a local patch at Isilon
was incorrectly managing numfreebufs for a new buf state.

Requested by:	Kostic Belousov


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@212562 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f