MFC r212359:
Fix a problem where device detection would work unreliably on Serverworks
K2 SATA controllers. The chip's status register must be read first, and
as a long, for other registers to be correctly updated after a command, and
this includes the command sequence in device detection as well as the
previously handled case after interrupts. While here, clean up some
previous hacks related to this controller.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@213291 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f