MFC: r213873
Explicitly lower the PIL to 0 as part of enabling interrupts, similar to
what is done on other platforms. Unlike as with the sched_throw(NULL)
called on BSPs during their startup apparently there's nothing which will
reliably lower it on APs. I'm unsure why this only came up on V215 though,
breaking these with r207248. My best guess is that these are the only
supported ones so far fast enough to lose some race.

PR:		151404


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@213953 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f