MFC r214854:
Validate whether the zfs_cmd_t submitted from userland is not smaller than
what we have.  Without the check the kernel could accessing memory that
does not belong to the request struct.

Note that we do not test if the struct equals in size at this time, which
may faciliate forward compatibility with newer binaries.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@215155 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f