MFC: r176965
Any driver that relies on its parent to set the devclass has no way to
know if has siblings that need an actual probe.  Introduce a specail
return value called BUS_PROBE_NOOWILDCARD.  If the driver returns
this, the probe is only successful for devices that have had a
specific devclass set for them.

Reviewed by: current@, jhb@, grehan@


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@215537 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f