MFC: r216820 (partial)
Add CPUTYPE support for sparc64. The net result is that it's now possible
to let the compiler optimize for the famility of UltraSPARC-III CPUs as the
default already was to optimize for UltraSPARC-I/II and generating generic
64-bit V9 is mainly for reference purposes. At least for SPARC64-V CPUs
code optimized for UltraSPARC-I/II still is the most performant one.
Thanks go to Michael Moll for testing SPARC64-V.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@217005 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f