MFC r215687:
Use g_eventlock to protect against losing wakeups in the g_event process
and replace tsleep(9) with msleep(9) which doesn't use a timeout. The
previously used timeout caused the event process to wake up ten times
per second on an idle system.

one_event() is now called with the topology lock held and it returns
with both the topology and event locks held when there are no more
events in the queue.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@219378 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f