use even larger stack size for ZFS txg_sync_thread
While the stack size was larger than the default stack size on i386, it
was smaller than the default stack size on amd64 and apparently that
wasn't enough.  So, bump the size to 4 pages.  Upcoming ZFSv28 code uses
8 pages for this stack size.

This is a direct commit to stable/8.

PR:		kern/154681
Discussed with:	pjd


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@219526 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f