MFC of 198254, 198255, 198350, 198267, 209190, 208831, 208830, 210243
198254:
When tzsetup is run as non-root and the "CMOS clock question on
UTC" is answered as No, it would abort without properly ending the
dialog session.

198255:
Make the usage of the default zoneinfo file to install clearer.

198350:
- Add support for chrooted installs.
- Add examples to the man-page.

198267:
Instead of having to know which timezone was picked last time, you
now can run "tzsetup -r" which will reinstall the last choice. This
data is recorded in /var/db/zoneinfo.

209190:
Use literal format strings.  Found by clang.

208831:
Add comment that this value is unused.
It is obvious that it isn't used, but both clang and Coverity talk about it.

208830:
When there is a problem with writing, also bail out.

Found with the clang checker.

210243:
Fix support for chrooted installs.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@220183 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f