MFC r221632, r221643:
r221632:

Fix isitme(), which is used to check if node-specific configuration
belongs to our node, and was returning false positive if the first
part of a node name matches short hostname.

r221643 (pjd):

Allow to specify remote as 'none' again which was broken by r219351, where
'none' was defined as a value for checksum.

Reported by:    trasz

Approved by:	pjd (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@222138 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f