MFC r223227: rc.subr: Eliminate about 100 forks from the boot sequence.
With the current sh, placing eval in a command substitution always results
in a fork(), even if it is the only command and only executes a single
simple command. Therefore, avoid it where it can be avoided easily.

Side effect: values starting with a hyphen and all whitespace are preserved.
The values are defaults and names for rc.conf variables and messages to be
given about obsolete ones.

The change in the _echoonce function is not included in this MFC because
stable/8 does not have this function.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@223742 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f