MFC: r223126 (partial)
Don't include curcpu in the mask which is used as the IPI cookie as we
have to ignore it when sending the IPI anyway. Actually I can't think of
a good reason why this ever was done that way in the first place as it's
not even usefull for debugging.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@224356 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f