MFC r222932: skel/.shrc: Improve commented CDPATH example for POSIX requirements.
POSIX says an empty entry in CDPATH shall not result in the new directory
being printed, while any non-empty entry shall result in the new directory
being printed, including ".". Therefore, the value of CDPATH should almost
always start with a colon, not dot and colon.

Our sh does not print the name for empty entries as well as "." entries.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@224409 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f