MFC r224865: tail: Fix crash if -F'ed file's filesystem disappears.
If tail notices that a file it is following no longer exists (because stat()
fails), it will output any final lines and then close the file. If the read
operation also causes an error, such as when the filesystem is forcefully
unmounted, it closes the file as well, leading to fclose(NULL) and a
segmentation fault.

PR:		bin/159750


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@225068 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f