MFC 226224
All indirect routes will fail the rtcheck, except for a special host
route where the destination IP and the gateway IP is the same. This
special case handling is only meant for backward compatibility reason.
The last commit introduced a bug in the route check logic, where a
valid special case is treated as an error. This patch fixes that bug
along with some code cleanup.

Suggested by:	glebius smirnoff
Reviewed by:	kmacy, discussed with glebius smirnoff


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@226330 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f