MFC 226713
Exclude host routes when checking for prefix coverage on multiple
interfaces. A host route has a NULL mask so check for that condition.
I have also been told by developers who customize the packet output
path with direct manipulation of the route entry (or the outgoing
interface to be specific). This patch checks for the route mask
explicitly to make sure custom code will not panic.

PR:		kern/161805


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@226877 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f