MFC 226710
The host-id/interface-id can have a specific value and is properly
masked out when adding a prefix route through the "route" command.
However, when deleting the route, simply changing the command keyword
from "add" to "delete" does not work. The failoure is observed in
both IPv4 and IPv6 route insertion. The patch makes the route command
behavior consistent between the "add" and the "delete" operation.

Approved by:	re (kib)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@227008 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f