MFC r226898:
Fix the text and attribution of the poem "Youth" by Samuel Ullman.

Our copy was a bastardization of the "older" (probably original?) and
shorter version of the poem; and the "newer" version that is better
known, and arguably more popular. Standardize on the latter.

Cf. http://www.bartleby.com/73/2099.html

Approved by:	re (kensmith)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@227133 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f