MFC 227460
A default route learned from the RAs could be deleted manually
after its installation. This removal may be accidental and can
prevent the default route from being installed in the future if
the associated default router has the best preference. The cause
is the lack of status update in the default router on the state
of its route installation in the kernel FIB. This patch fixes
the described problem.

Reviewed by:	hrs, discussed with hrs


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@228092 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f