MFC 225518,225793,227085: Allow the ipfw.ko module built with a kernel to honor any options defined in the kernel config.  This more closely matches the behavior of other modules which inherit configuration settings from the kernel configuration during a kernel + modules build.
Do not try to build the module in case of no INET support but keep #error
calls for now in case we would compile it into the kernel.

While here garbage collect unneeded opt_*.h includes.
opt_ipdn.h is not used anywhere but we need to leave the DUMMYNET
entry in options for conditional inclusion in kernel so keep the
file with the same name.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@229497 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f