MFC r228971:
For sys/dev/ce/tau32-ddk.c, disable the following warning when building
with clang:

sys/dev/ce/tau32-ddk.c:1228:37: warning: implicit truncation from 'int' to bitfield changes value from 65532 to 8188 [-Wconstant-conversion]

Since this file is obfuscated C, we can never determine (in a sane way,
at least :) if this points to a real problem or not.  The driver has
been in the tree for more than five years, so it most likely isn't.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@229602 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f