MFC r228978:
For several files in sys/dev/drm, disable -Wunused-value when building
with clang.  There are several macros in these files that return values,
and in some cases nothing is done with them, but it is completely
harmless.  For some other files, also disable -Wconstant-conversion,
since that triggers a false positive with the DMA_BIT_MASK() macro.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@229731 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f