MFC 230877: Fix NULL dereference panic on attempt to turn off (on system shutdown) disconnected swap device.
This is quick and imperfect solution, as swap device will still be opened
and GEOM will not be able to destroy it. Proper solution would be to
automatically turn off and close disconnected swap device, but with existing
code it will cause panic if there is at least one page on device, even if
it is unimportant page of the user-level process. It needs some work.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@231188 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f