MFC:	r231168
Update the data structures with some fields reserved for
ext4 but that can be used in ext3 mode.

Also adjust the internal inode to carry the birthtime,
like in UFS, which is starting to get some use when
big inodes are available.

This is a prerequisite for new features that will be
coming in the future.

Approved by:	jhb (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@231528 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f