MFC r232267 via stable/8 MFC r232354 Workaround for PCIe 4GB boundary issue
Enforce a boundary of no more than 4GB - transfers crossing a 4GB
boundary can lead to data corruption due to PCIe limitations.  This
change is a less-intrusive workaround that can be quickly merged back
to older branches; a cleaner implementation will arrive in HEAD later
but may require KPI changes.

This change is based on a suggestion by jhb@.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@232397 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f