MFC 227470:
Report the amount of memory from smbios data if provided.

This should get the correct memory size even if a 32-bit image is running
on a machine with > 4GB of memory.  This can be useful is using a 32-bit
installer on a machine which will eventually run a 64-bit image.

Reviewed by:	kmoore


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@233217 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f