MFC r228648:
On start most of sysctl_kern_proc functions use the same pattern:
locate a process calling pfind() and do some additional checks like
p_candebug(). To reduce this code duplication a new function pget() is
introduced and used.

As the function may be useful not only in kern_proc.c it is in the
kernel name space.

Suggested by:	kib
Reviewed by:	kib


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@233926 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f