MFC r233713 Remove the magic mfi_array is 288 bytes and just use the sizeof the array since it is not 288 bytes.
Change reporting of a "SYSTEM" disk to "JBOD" to match
LSI MegaCli and firmware reporting.

This means that mfiutil command to "create jbod" is now a
little confusing since a RAID per drive is not really what
LSI defines JBOD to be.  This should be fixed in the future
and support added to really create LSI JBOD and enable that
feature on cards that support it.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@234433 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f