MFC r234775:
As cron(8) is started with '-s' by default, timezones that observe
DST should not need to worry about scheduling jobs when the DST time
changes.

Rather than removing the BUGS section in crontab(5) regarding this,
note that disabling '-s' may still cause jobs to be executed twice or
not at all.

PR:		166318


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/7@235049 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f