MFC r237348:
Currently the code uses gzFile * for a zlib file descriptor, which
is not correct.  The code works by accident because gzFile is
currently defined as void *, and internally it would be casted from
or to its real type.

A newer version of zlib will instead define it as a pointer to a
specific type pointer (namely, struct gzFile_s *).  This therefore
would cause stricter checks and compiler would catch this type
mismatch.

This change does not cause any changes to the resulting binary,
as validated with md5(1).


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@237464 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f