MFC: r236581
The loaddev environment variable is not modifiable once set, so it is not
update for ZFS. It seems that this does not really affect anything except
the help command. Nevertheless, rearrange things so loaddev is set only
once in all cases in order to get it right.
Pointed out by: avg


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@237815 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f