MFS: r239148
    r239148 | imp | 2012-08-08 14:21:33 -0600 (Wed, 08 Aug 2012) | 8 lines

    Make the addition of the d_gone binary compatible.  This allows
    storage drivers compiled for 9.0 to work on 9.1 and preserves the ABI
    for disks.

    Reviewed by:    scottl, ken
    Approved by:    re@ (kensmith)
    MFS after:      releng/9.1 gets sorted

This change fixes the binary compatibility problems with additions to
the disk(9) API.  Disk drivers compiled against 9.0 will not work on
9.1-RC1, but should start working after this change or in 9.1-RC2 and
later.  Disk drivers should not be linked against 9.1-RC1 for
distribution, except specifically for testing on RC1.  This was
planned for RC1, but it was delayed due to circumstancs beyond my
control.

Approved by: re@ (kensmith)


git-svn-id: svn+ssh://svn.freebsd.org/base/releng/9.1@239707 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f