Very belated MFC of 227548
Was chasing down a failure to load f/w on a 2400. It turns out that the card
is actually broken, or needs a BIOS upgrade for 64 bit loads, but this uncovered
a couple of misplaced opcode definitions and some missing continual mbox command
cases, so might as well update them here.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@240008 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f