MFC r242625:
Remove duplicate const specifiers in many drivers (I hope I got all of
them, please let me know if not).  Most of these are of the form:

static const struct bzzt_type {
      [...list of members...]
} const bzzt_devs[] = {
      [...list of initializers...]
};

The second const is unnecessary, as arrays cannot be modified anyway,
and if the elements are const, the whole thing is const automatically
(e.g. it is placed in .rodata).

I have verified this does not change the binary output of a full kernel
build (except for build timestamps embedded in the object files).

Reviewed by:	yongari, marius


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@242908 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f