MFC r242666, r242671.
r242666:
Remove the tid from the software table (and bump down the in-use
counter) when the syncache doesn't want the driver to reply to an
incoming SYN.  This fixes a harmless bug where tids_in_use would
go out of sync with the hardware counter.

r242671:
Make sure the inp hasn't been dropped before trying to access its socket
and tcpcb.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@243098 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f