MFC r244557:
Add libnetbsd, a thin compatibility layer intended to allow a limited
set of NetBSD software to compile as part of the FreeBSD build with
little or no modification.  It is built as a static library and not
installed for general use.  Likewise, its header files are not
installed.

Sponsored by:	DARPA, AFRL


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@245461 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f