partial MFC of rev=245362: enable building virtio devices into static kernels.
I think the 'files.*' entries should be improved (also in HEAD) because
bringing up a vtnet device now requires 3 entries in your kernel config:
"device virtio, device virtio_pci, device vtnet"

but i'll leave the fix to a future commit.

This is also the reason not to enable the device in GENERIC kernels now.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@246885 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f