MFC r246522: sh: Simplify mksyntax and make it fit for cross-compiling.
Now it outputs fixed files, which use constants provided by the C standard
library to determine appropriate values for the target machine.

Before, mksyntax inspected the host machine which resulted in subtle
breakage if e.g. char is signed on the host and unsigned on the target such
as when cross-compiling on x86 for ARM.

Submitted by:	Christoph Mallon


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@247163 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f