MFC r249047 (avg): spa_open_common: fix argument to zvol_create_minors
Prior to r248571 spa_open was always called with a bare pool name,
but now it is called with a dataset name instead (spa_lookup handles
that).
So, when a ZFS root is mounted spa_open is called with a name of a root
dataset, which can very well be different from the pool name.
But zvol_create_minors should be called with the pool name, because it
performs a recursive traversal of all datasets under the name to find
all those that are volumes.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@249689 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f