MFC r250337
Work around the implementation of LIBRARIES_ONLY.  It causes drti.o to
not be installed in /usr/lib32 on systems with compat-32 support.

This fix has two parts.  First, the build is forced by linking drti.o
into a dummy internal library.  Second, the object file is installed
manually in the LIBRARIES_ONLY case.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@250534 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f