MFC of 250708:
Clean up trailing whitespace.

Submitted by: Andy Kosela

MFC of 250710:

When running the -m option to generate a newfs(8) command suitable for
recreating the filesystem, check for and output the -i, -k, and -l
options if appropriate.

Note the remaining deficiencies of the -m option in the dumpfs(8)
manual page. Specifically that newfs(8) options -E, -R, -S, and -T
options are not handled and that -p is not useful so is omitted.
Also document that newfs(8) options -n and -r are neither checked
for nor output but should be. The -r flag is needed if the filesystem
uses gjournal(8).

PR:           bin/163992
Reported by:  Dieter <freebsd@sopwith.solgatos.com>
Submitted by: Andy Kosela <akosela@andykosela.com>


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@251043 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f