MFC r248282, in a modified fashion.  From the original changelog:
	Add currently unused flag argument to the cluster_read(),
	cluster_write() and cluster_wbuild() functions.  The flags to be
	allowed are a subset of the GB_* flags for getblk().

This merge adds a cluster_*_gb() API variant instead of changing the ABI
with an added argument to the existing API.  Most API consumers that were
changed in the original rev have been left un-changed in this merge to
reduce churn.  The mergeinfo is recorded though for future merging
convenience.  This is effectively a no-op for the moment.

Submitted by:	kib, FF
Obtained from:	Netflix


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@251144 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f