MFC: r238172 (partial)
Don't use Maxmem when the amount of memory is meant. Use realmem instead.
Maxmem is not only a MD variable, it represents the highest physical memory
address in use. On systems where memory is sparsely layed-out the highest
memory address and the amount of memory are not interchangeable. Scaling the
AGP aperture based on the actual amount of memory (= realmem) rather than
the available memory (= physmem) makes sure there's consistent behaviour
across architectures.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@251994 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f