MFC r252504:
Import an implementation of the CAIA Delay-Gradient (CDG) congestion control
algorithm, which is based on the 2011 v0.1 patch release and described in the
paper "Revisiting TCP Congestion Control using Delay Gradients" by David Hayes
and Grenville Armitage. It is implemented as a kernel module compatible with the
modular congestion control framework.

CDG is a hybrid congestion control algorithm which reacts to both packet loss
and inferred queuing delay. It attempts to operate as a delay-based algorithm
where possible, but utilises heuristics to detect loss-based TCP cross traffic
and will compete effectively as required. CDG is therefore incrementally
deployable and suitable for use on shared networks.

In collaboration with:	David Hayes <david.hayes at ieee.org> and
				Grenville Armitage <garmitage at swin edu au>
Sponsored by:	Cisco University Research Program and FreeBSD Foundation


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@252952 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f