MFC	r253050:
Enhancement when writing an entire block of a file (from UFS r231313).

This change first attempts the uiomove() to the newly allocated
(and dirty) buffer and only zeros it if the uiomove() fails. The
effect is to eliminate the gratuitous zeroing of the buffer in
the usual case where the uiomove() successfully fills it.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@253218 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f