MFC r240966 (by brooks)
Add a COMPILER_FEATURES variable that is set in bsd.compiler.mk.

When the compiler is clang add c++11 to COMPILER_FEATURES.

When the compiler supports c++11, default to building libc++.  This will
not impact the compliation of programs, but will make it easier for
users with clang compiled systems to test libc++ by passing -stdlib=libc++
to their c++ compliations.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@253220 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f