MFC of 253330
When fiddling with options of which registers to copy out for
a mailbox command and which registers to copy back in when
the command completes, the bits being set need to not only
specify what bits you want to add from the default from the
table but also what bits you want *subtract* (mask) from the
default from the table.

A failing ISP2200 command pointed this out.

Approved by:	re


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@253371 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f