MFC 254372:
Export the maxio field in the CAM XPT_PATH_INQ CCB in the isp(4)
driver.

This tells consumers up the stack the maximum I/O size that the
controller can handle.

The I/O size is bounded by the number of scatter/gather segments
the controller can handle and the page size.  For an amd64 system,
it works out to around 5MB.

Reviewed by:	mjacob
Sponsored by:	Spectra Logic


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@254971 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f