MFC r255775 r255796 r255807 r255930 r255929 r255957:
Create a separate script to generate osreldate.h rather than sourcing
newvers.sh into a temporary subshell with inline make rules.
Fixes PR 160646.

Allow the path to the system source directory to be passed in to
newvers.sh.  Pass it in from include/Makefile.  If it isn't passed in,
fall back to the old logic of using dirname $0.  Fixes PR 174422.

PR:		160646 174422


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@256386 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f