MFC: r257862, r257863, r257864
r257862:

Use explicit long cast to avoid overflow in bitopts.

This was causing problems with the buddy allocator inside of
ofed.

r257863:

Fix for bad performance when mtu is increased.

Update the auto moderation behavior in the mlxen driver to match
the new LINUX OFED code.

r257864:

Do not use a sleep lock when protecting the driver flags.

This was causing a locking issue with lagg.

Approved by:	re


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@257867 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f