Fix an inappropriate free of a non-dynamic value. While here, make the code more naive and robust: 1.  When setting ev_value, also always set ev_flags appropriately 2.  Always check ev_value and ev_flags before calling free.
Both the value and the EV_DYNAMIC property can come directly from the
consumers of the environment functionality, so it's good to be careful.
And since this code is typically not looked at for long periods of
time, it's good to have it be a little "dumb-looking".

Trigger case for the bug:
        env_setenv("foo", 0, "1", NULL, NULL);
        env_setenv("foo", 0, "2", NULL, NULL);

Obtained from:	Juniper Networks, Inc.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@259561 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f