MFC r259527:
Do not create a hardware IPv6 server if the listen address is not
in6addr_any and is not in the CLIP table either.  This fixes a reported
TOE+IPv6 NULL-dereference panic in do_pass_open_rpl().

While here, stop creating hardware servers for any loopback address.
It's just a waste of server tids.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@259804 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f