MFC	r260988, r261034, r261120,  r261235:
ext2fs: Properly the EXT4_EXTENTS and EXT4_INDEX to the inode flags.

In order to support Ext4 extents we need to pass the Ext4 inode flags
without interfering with the chflags. This is better done by using the
i_flag field in the inode and doing proper translation to the linux
ext4 equivalents.

Solve a potential corruption issue in the dirindex code. The dirindex
code can now be renabled as the problems related to it have been
solved.

Suggested by:	bde
Tested by:	kevlo


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@261312 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f