MFC: r261655
Refer newfs and growfs users to fsck_ffs instead of
fsck, the latter does not accept the referred to "-b" flag.

The part of the change pertaining newfs has already been merged
to stable/9 in r237505, so only record that part of the merge.

PR:		82720
Submitted by:	David D.W. Downey


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@261964 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f