MFC r249514 (by gnn): Point args[0] not at the thread that is ending but at the one that is starting.  This is in line with practice in OpenSolaris.
Note that this change is only in ULE and not in the 4BSD scheduler.
Once this change settles in (MFC timeout has expired) we'll try it out
on 4BSD as well.

MFC r260043:
The arguments to sched:::off-cpu are the thread and associated process of
the thread selected to run, not the currently running thread.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@262012 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f