MFC r257650:
Defer loading of kernel and modules if the beastie menu is enabled. Add a
kernel selection menu to the beastie menu. List of kernels is taken from
`kernels' in loader.conf(5) as a space (or comma) separated list of names
to display (up to 9). If not set, default value is "kernel kernel.old".
Does not validate that kernels exist because the next enhancement will be
to allow selection of the root device.

Discussed on:	-current


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@262701 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f