MFC r260046 (by glebius):
Fix the parse type for NGM_LISTTYPES.

Actually, text versions of generic commands are not used, since ngctl(8)
uses binary messages for them. And to request a text command one needs
a working ngctl(8). That's why the bug was never discovered. I'm pondering
on removing the text support for generic commands.

Found by:	dim with clang 3.4


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@263452 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f