MFC: r263121
Re-format the license to conform to our BSD license template as much
as possible.  This does not change the wording in any way.

Remove the 3rd clause ("advertising clause") of the BSD license as
permitted by the University of Berkeley on July 22, 1999.  While the
clause itself mentions Lawrence Berkeley Laboratory, UCB is the sole
copyright holder of this file.

Reviewed by:	imp, emaste, eadler


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@263830 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f