MFC r261558
Add a new sysctl, dev.cxgbe.N.rsrv_noflow, and a companion tunable,
hw.cxgbe.rsrv_noflow.  When set, queue 0 of the port is reserved for
TX packets without a flowid.  The hash value of packets with a flowid
is bumped up by 1.  The intent is to provide a private queue for
link-level packets like LACP that is unlikely to overflow or suffer
deep queue latency.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@264493 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f