MFC r259382:
Read card capabilities after firmware initialization, instead of setting
them up as part of firmware initialization (which the driver gets to do
only if it's the master driver).

Read the range of tids available for the ETHOFLD functionality if it's
enabled.

New is_ftid() and is_etid() functions to test whether a tid falls within
the range of filter tids or ETHOFLD tids respectively.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@265426 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f