MFC r254248, r254670:
Make the features a 64-bit value instead of 32-bit.
This will allow an easier integration of the support
for NDATA.
While there, do also some minor cleanups.
Obtained from:	rrs@

Make also the features of the association 64-bit.
When exporting to xinpcb, just export the lower
32-bit. Using there also 64-bits will break the
ABI and will be committed separetly.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/9@265958 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f