MFC r257549, r261642
Don't create a distinct free page pool for segregating allocations that are
accessed through the direct map unless the kernel configuration actually
includes a direct map.  Only a few configurations do, and for the rest the
unnecessary free page pool is a small pessimization.

Remove the ARM_USE_SMALL_ALLOC option and code related to it.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@266175 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f