MFC r266426
Slightly restructure the final loop in rman_reserve_resource_bound().
Replace with the existing loop termination test with a similar
condition from the nested "if" that may terminate the loop a bit
sooner, but still not too early.   This condition can then be removed
from the nested "if".  Relocate an operator to be style(9) compliant.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@266521 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f