Remove the unnecessary i2c slave address assignment.
The ti_i2c controller only works in the master mode and the i2c address
passed on iicbus_reset() is used to set the controller slave address when
operating as an i2c slave (which isn't currently supported).

When talking to a slave, the slave address is correctly provided to
ti_i2c_tranfer().


git-svn-id: svn+ssh://svn.freebsd.org/base/head@267009 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f