MFC changes related to PR kern/189089.  Unlike CURRENT, stable/10 does not panic when you attempt to remove the IP address.  But it still fails to remove the address.
MFC r265094

Add regression test for PR kern/189088.

MFC r265092

Fix a panic when removing an IP address from an interface, if the same
address exists on another interface.  The panic was introduced by change
264887, which changed the fibnum parameter in the call to rtalloc1_fib() in
ifa_switch_loopback_route() from RT_DEFAULT_FIB to RT_ALL_FIBS.  The
solution is to use the interface fib in that call.  For the majority of
users, that will be equivalent to the legacy behavior.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@267195 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f