MFC r267674:
Honor jails for unbound SCTP sockets when selecting source addresses,
reporting IP-addresses to the peer during the handshake, adding
addresses to the host, reporting the addresses via the sysctl
interface (used by netstat, for example) and reporting the
addresses to the application via socket options.
This issue was reported by Bernd Walter.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@267769 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f