MFC	r268066: regex(3): Add support for \< and \> word delimiters
Solaris and other OSs have support for \< and \> as word
delimiters in utilities like sed(1). These are useful to
have for general compatiblity with Solaris but should be
avoided for portability with other systems, including the
traditional BSDs.

Bump __FreeBSD_version as this is likely to affect some
userland utilities.

Reference:
https://www.illumos.org/issues/516

PR:		bin/153257
Obtained from:	Illumos


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@269484 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f