MFC	r269662:
10 has a new flex (2.5.37) and the config.h for unbound has been updated to
take this into account. Alas it breaks source upgrade from any version of
9 because flex is not built as a bootstrap-tools (it would be for older
versions).

That means "libunbound/configlexer.c" is built with the old flex but using
config.h for the new one. Build is thus broken going from 9.* to 10.

Make flex a bootstrap-tools entry if host is less than 1000033 to take into
account the flex update in 10.

Tested on both 9.2-RC3 and 9.3 by myself and dim@. Running buildworld in
head but as both 10 and 11 has the new flex, it will not matter.

Reviewed by:    imp
Approved by:    des, imp
MFC after:      1 week
Phabric:        D554


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@269967 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f