MFC r270976: Allow standalone debug for non-default ${PROG} targets
This allows WITH_DEBUG_FILES to produce standalone debug for the ELF
runtime linker.

We previously disabled standalone debug files for bsd.prog.mk consumers
that included a non-default ${PROG} target, but this is not required.

Consumers that do not support standalone debug are still handled by
disabling it for statically linked binaries, and for those that specify
a non-default binary format.

Sponsored by:	DARPA, AFRL
Approved by:	re


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@271274 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f