MFC r234098 (jhb) and r255047.
r234098:
Add media types for 40G media that might be used with FreeBSD.

r255047:
Add a routine for attaching an mbuf to a buffer with an external
refcount.  This one is willing to work with buffers that may already be
referenced.  MEXTADD/m_extadd are suitable only for the first attachment
to a cluster -- they initialize the refcount to 1.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/8@271547 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f