xen/balloon: fix accounting of current memory pages on PVH
Using realmem on PVH is not realiable, since in this case the realmem value
is computed from Maxmem, which contains the higher memory address found. Use
HYPERVISOR_start_info->nr_pages instead, which is set by the hypervisor and
contains the exact number of memory pages assigned to the domain.

Sponsored by: Citrix Systems R&D


git-svn-id: svn+ssh://svn.freebsd.org/base/head@272320 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f