MFC r271784 - Fix the handling of EOP in status descriptors for if_igb(4) and don't double-free mbufs.
Like ixgbe(4) chipsets, EOP is only set on the final descriptor
in a chain of descriptors.  So, to free the whole list of descriptors,
we should free the current slot _and_ the assembled list of descriptors
that make up the fragment list.

The existing code was setting discard once it saw EOP + an error status;
it then freed all the subsequent descriptors until the next EOP. That's
totally the wrong order.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@272969 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f