MFC r272627: Checksum coverage values larger than 65535 for UDPLite are invalid. Check for this when the user calls setsockopt using UDPLITE_{SEND,RECV}CSCOV.
MFC r272628:
When plen != ulen, it should only be checked when this is UDP.

MFC r272645:
If the checksum coverage field in the UDPLITE header is the length
of the complete UDPLITE packet, the packet has full checksum coverage.
So fix the condition.

MFC r272660:
UDPLite requires a checksum. Therefore, discard a received packet if
the checksum is 0.

MFC r272661:
The default for UDPLITE_RECV_CSCOV is zero. RFC 3828 recommend
that this means full checksum coverage for received packets.
If an application is willing to accept packets with partial
coverage, it is expected to use the socket option and provide
the minimum coverage it accepts.

MFC r272662:
Fix the checksum computation for UDPLite/IPv6. This requires the
usage of a function computing the checksum only over a part of the function.
Therefore introduce in6_cksum_partial() and implement in6_cksum() based
on that.
While there, ensure that the UDPLite packet contains at least enough bytes
to contain the header.

MFC r272663:
Check for UDP/IPv6 packets that the length in the UDP header is at least
the minimum. Make the check similar to the one for UDPLite/IPv6.

MFC r272664:
UDP/IPv6 and UDPLite/IPv6 require a checksum. So check for it.

MFC r272754:
Fix a bug introduced in
https://svnweb.freebsd.org/base?view=revision&revision=272347

Approved by: re (gjb)


git-svn-id: svn+ssh://svn.freebsd.org/base/releng/10.1@272991 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f