MFV r274273:
ZFS large block support.

Please note that booting from datasets that have recordsize greater
than 128KB is not supported (but it's Okay to enable the feature on
the pool).  This *may* remain unchanged because of memory constraint.

Limited safety belt is provided for mounted root filesystem but use
caution is advised.

Illumos issue:
    5027 zfs large block support

MFC after:	1 month


git-svn-id: svn+ssh://svn.freebsd.org/base/head@274337 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f