MFC 275415:
Allow the UAL APSR_nzcv format for the mrc and mrc2 instructions. The clang
integrated assembler only allows these forms so binutils will need to
support them.

Sponsored by:	ABT Systems Ltd

MFC 275584:

Fix mrc and mrc2 with APSR_nzcv. Binutils encodes it internally as 0 where
we need it to be set to 15 for it to be equivalent to r15.

Sponsored by:	ABT Systems Ltd


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@275770 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f