MFC r275481: Add to CTL support for threshold notifications for file-backed LUNs.
Previously it was supported only for ZVOL-backed LUNs, but now should work
for file-backed LUNs too.  Used value in this case is a space occupied by
the backing file, while available value is an available space on file
system.  Pool thresholds are still not implemented in this case.

Sponsored by:   iXsystems, Inc.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@275893 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f