MFC r273610:
Add an iicbus_reset() method to bcm2835_bsc.  While it is generally not
used for kernel devices it is used by i2c(8).

This fix the 'error: Device not configured' when i2c(8) tries to reset the
controller:

# i2c -r
Resetting I2C controller on /dev/iic0: error: Device not configured

For now use conservative settings for default i2c speeds.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@276869 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f