Merge 272659:
Add netmap support to libpcap. Tcpdump and other native pcap clients
can now run directly on netmap ports using netmap:foo or valeXX:YY
as device names.

Modifications to existing code are small and trivial,
the netmap-specific code is all in a new file.

Please be aware that in netmap mode the physical interface is
disconnected from the host stack, so libpcap will steal the traffic
not just make a copy.

For the full version of the code (including linux and autotools support) see
	https://code.google.com/p/netmap-libpcap/


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@277722 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f