MFC r271025, r271029, r271030 (by sbruno):
MFV:  Only emit movw on ARMv6T2

Building for the FreeBSD default target ARMv6 was emitting movw ASM on certain
test cases (found building qmake4/5 for ARM).  Don't do that, moreover, the AS
in base doesn't understand this instruction for this target.  One would need
to use --integrated-as to get this to build if desired.

http://llvm.org/viewvc/llvm-project?view=revision&revision=216989

Submitted by:	ian
Reviewed by:	dim
Obtained from:	llvm.org


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@278709 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f