MFC r260571: fts: Stat things relative to the directory fd, if possible.
As a result, the kernel needs to process shorter pathnames if fts is not
changing directories (if fts follows symlinks (-L option to utilities), fts
cannot open "." or FTS_NOCHDIR was specified).

Side effect: If pathnames exceed PATH_MAX, [ENAMETOOLONG] is not hit at the
stat stage but later (opendir or application fts_accpath) or not at all.


git-svn-id: svn+ssh://svn.freebsd.org/base/stable/10@281711 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f